test/e2e/requirements.txt
PyYAML==5.3.1
requests==2.26.0
apache-skywalking==0.6.0
Flask==2.0.2